Update order-status.md
I have updated errors in the text and deleted Scenario F Over Shipment.

If some reels are taken from stock, then they are normally included in Produced quantity. It would mean that Produced and Shipped quantities would be the same when ShipmentCompleted, so this isn't really a new scenario.

We also have other scenarios, e.g. the products can be invoiced before they are shipped. We have a major difference in this scenario compared to existing scenarios, orderStatus=Completed is after ShipmentCompleted and not after the issue of the invoice as in existing scenarios.
How do we communicate the status that the order line items are invoiced when invoicing is done before shipments? Or is it enough to report an Invoiced quantity?
Should we add this scenario?

We should also discuss if we need more quantity types. E.g. a merchant is selling sheeted paper from stock, would he use quantity type Produced in the API response, when he has allocated stock items to the customer order? The merchant is the supplier in this case.