Update shipment-status.md
Changed shipmentArrivalStatus from OnTime to Delayed for shipment SHP002 because estimatedDateTimeOfArrival is later than scheduled time.
Change shipmentNumber in scenario C to be unique by the supplier. The Supplier can use many forwarders so shipmentNumber of the supplier has to be unique.
Changed some errors in the text.

N.B. I have not updated Mock files that have to updated accordingly.